Kangaroo
There are two kangaroos on a number line ready to jump in the positive direction (i.e, toward ). Each kangaroo takes the same amount of time to make a jump, regardless of distance. That is, if kangaroo one jumps 3 meters and kangaroo two jumps 5 meters, they each do it in one second, for example.

Given the starting locations and jump distances for each kangaroo, determine if and when they will land at the same location at the same time.

Input Format

    4 space-separated integers
        : starting locations and meters per jump for kangaroos and

Output Format

Print YES if they can land on the same location at the same time. Otherwise, print NO.